Remove superfluous cast

+++ libctru/source/services/cfgu.c
@@ -351,8 +351,8 @@ Result CFGI_GetLocalFriendCodeSeedData(u8 *data)
 	u32 *cmdbuf = getThreadCommandBuffer();
 
 	cmdbuf[0] = IPC_MakeHeader(0x404,1,2); // 0x4040042
-	cmdbuf[1] = (u32)0x110;
-	cmdbuf[2] = IPC_Desc_Buffer((u32)0x110, IPC_BUFFER_W);
+	cmdbuf[1] = 0x110;
+	cmdbuf[2] = IPC_Desc_Buffer(0x110, IPC_BUFFER_W);
 	cmdbuf[3] = (u32)data;
 
 	if(R_FAILED(ret = svcSendSyncRequest(cfguHandle)))return ret;
@@ -380,8 +380,8 @@ Result CFGI_GetSecureInfoData(u8 *data)
 	u32 *cmdbuf = getThreadCommandBuffer();
 
 	cmdbuf[0] = IPC_MakeHeader(0x814,1,2); // 0x8140042
-	cmdbuf[1] = (u32)0x11;
-	cmdbuf[2] = IPC_Desc_Buffer((u32)0x11, IPC_BUFFER_W);
+	cmdbuf[1] = 0x11;
+	cmdbuf[2] = IPC_Desc_Buffer(0x11, IPC_BUFFER_W);
 	cmdbuf[3] = (u32)data;
 
 	if(R_FAILED(ret = svcSendSyncRequest(cfguHandle)))return ret;
@@ -395,8 +395,8 @@ Result CFGI_GetSecureInfoSignature(u8 *data)
 	u32 *cmdbuf = getThreadCommandBuffer();
 
 	cmdbuf[0] = IPC_MakeHeader(0x815,1,2); // 0x8150042
-	cmdbuf[1] = (u32)0x100;
-	cmdbuf[2] = IPC_Desc_Buffer((u32)0x100, IPC_BUFFER_W);
+	cmdbuf[1] = 0x100;
+	cmdbuf[2] = IPC_Desc_Buffer(0x100, IPC_BUFFER_W);
 	cmdbuf[3] = (u32)data;
 
 	if(R_FAILED(ret = svcSendSyncRequest(cfguHandle)))return ret;
